created validsnp method for genome input validation;created replace_coord_rsid for single snp input for ldpair and ldpop; created replace_coords_rsid_list for snp list inputs
